Ra master docker build n publish (#26)
GitAction which automates the build and deployment of the workflow-launcher-api and workflow-launcher-ui images. This required a change in the generation of URLs to clone for the boot script. Now the script attempts to pull from repos using the format git@aliasX.github.com:broadinstitute/repo-to-pull.git. If that doesn't work, then it reverts to the former way of using the format git@github.com:broadinstitute/repo-to-pull.git. This was required because the github action requires the use of unique deployment keys for the repos which are cloned.