Test for backwards compatibility?
Like, this can't actually be reached, but it does work and is the expected behavior. Not testing it would leave the WGS database changes untested. I guess I could remove the WGS database changes but I honestly think it is simpler to leave it, since the same loader is used by aou and copyfile