exception/wrap is useless
The "request" is actually the entire map of routes. So the big block of output that was the bug was actually the (logr/error request) line, and we don't have enough info in the wrapper anyway to figure out if the issue is server or client-side (so we know how much to print). In other words, this is unintuitive and inferior to just making our own exception handler log. Our own handler is the default so it will always be called.