Test that we now actually get 400 errors back
This is actually basically an integration test for the whole error-handling system because it checks that the :type is maintained, and then that the :type is parsed correctly, and then that the status code is maintained, and then that there are no further errors that would change it to a 500. This is concise but I think it covers a lot of ground without being too verbose or brittle.