Use batching requests for workflow statuses
Rather than passing `make-update-workflows` a function that will query Cromwell for the status of the workflow, we instead make a batch request beforehand for all non-terminal workflows and pass `make-update-workflows` a stub function that just parses the response.

This has the benefits of batching without locking us into that mechanism, and we can still share the underlying `make-update-workflows` with the update system for Terra.