[GH-1079] Fix WFL's exception logging (#264)
* Pass along ex-data

* exception/wrap is useless

The "request" is actually the entire map of routes. So the big block of output that was the bug was actually the (logr/error request) line, and we don't have enough info in the wrapper anyway to figure out if the issue is server or client-side (so we know how much to print). In other words, this is unintuitive and inferior to just making our own exception handler log. Our own handler is the default so it will always be called.

* Make our own handler log

* Move the middleware up to get pretty-printing of the response

* Test that we now actually get 400 errors back

This is actually basically an integration test for the whole error-handling system because it checks that the :type is maintained, and then that the :type is parsed correctly, and then that the status code is maintained, and then that there are no further errors that would change it to a 500. This is concise but I think it covers a lot of ground without being too verbose or brittle.

* Don't log handled errors

* Don't wrap when unnecessary

* fix test

* Split out logging into another handler

* Make the handler more concise

* Make Tom's suggestion

* rename and recomment