[GH-1096] Batch update (#269)
* Update query to work with the API

Seems like the API changed since this was used. I can't find any mentioning of pagination on Cromwell's documentation, so fingers crossed...

* Use batching requests for workflow statuses

Rather than passing `make-update-workflows` a function that will query Cromwell for the status of the workflow, we instead make a batch request beforehand for all non-terminal workflows and pass `make-update-workflows` a stub function that just parses the response.

This has the benefits of batching without locking us into that mechanism, and we can still share the underlying `make-update-workflows` with the update system for Terra.

* Remove now-unused function

This was a hack anyway, and we have no need for it since we can query the batch endpoint

* Very basic attempt at patching up the tests

* Revert "Update query to work with the API"

This reverts commit fd6572cc

* Get integration tests working

* remove println

* clean up empty case

* Replace wfl link to build.boot.

* Make Tom's suggestion

* Reset `query` back to how it was

* Change mock to work with querying for label

Since can't properly/easily mock Cromwell's response now (we aren't given information to do it easily), we instead just mock the updating of the statuses. This still covers the goal of the test here.

* Query via labels instead of id

A potentially significant performance hit over the wire because here we request for entire workloads, having no way to exclude workflows that we already know are finished.

If we ever hit this as a performance issue, reverting this commit, 0fea034c506471e2c2f4c1adefe6c3cc2b6630a3, and 04e332c6afe904c31e7d9356cfefa5bc9c86e30f will swap some things around to query by the ids that we are interested in.

* Make Ed's suggestions

Co-authored-by: Tom Lyons <tbl3rd@gmail.com>