[GH-1318] Stop Workload Source RR: https://broadinstitute.atlassian.net/browse/GH-1318 Add `stop-source` multimethod and implement for the TerraDataRepoSource. When stopped, the source will stop looking for new rows to snaphot in the TDR.
Also added logic for when the workload is finished. In this case, once
started the workload is finished when it has been stopped and both
source and executor queues are empty. To do this, I added another
multimethod to return the length of the queues.

For the executor, the length of the queue is the number of workflows
that have not been consumed or aborted.

For the source, the length of the queue is the number of unconsumed
records whose snapshot job is not 'failed'.

Note that this is slightly at odds with peek, where peek is meant to
return objects that are ready to be consumed.