[GH-1327] Fix Start After Stop (#411)
[GH-1327] Fix Start After Stop
In this change I'm fixing the strange behaviour were we can start a
covid workload after it was stopped. To do this, I've made it illegal to
stop a workload before it's been started.

I've done this because stop before start seems strange to me too. The
purpose of stop was to stop getting new data from the source but allow
all data in the workload pipeline to get flushed (ie. consume all
created snapshots and write all worklfows back). Stop was not meant for
aborting a workload and cancelling/aborting all active prossessing. If
there is a use case for such a thing (and i think there is) - I think we
should expose a new endpoint called `POST /api/v1/abort` or something.

Of course this PR is pretty small so if the team feels strongly about
stop before start I can find another solution/