[GH-1336] Fix continual snapshot creation (#419)
Our BigQuery query was wong as we were simply calling `toString` on a
java.sql.Timestamp. This lead to intervals like
`['2021-05-26 15:31:18.287641' '2021-05-26T20:00:13']`
The space was causing the problem and BigQuery interpreted that as a
date. The fix is to ensure the lower and upper bound of our query-between
interval is of the same type by coercing java.sql.Timestamp `last_checked`
into a java.time.OffsetDateTime.