[GH-1348] Remove :pipeline from COVID workload request (#426)
RR: https://broadinstitute.atlassian.net/browse/GH-1348
The :pipeline attribute of the workload request is meaningless for COVID-type workloads. Note that these workloads aren't even specific to COVID-19 processing, rather a generalisation.
In this PR, I've removed the :pipeline attribute. To do this, all workload operations on maps with a nil :pipeline will be implemented by the functions in the covid namespace.