[GH-1411] Describe Workflow Outputs (#471)
RR: https://broadinstitute.atlassian.net/browse/GH-1411
Inform the Sink downstream of the executor of the workflow type.

When sinking outputs to the TDR, you need to ingest the file outputs followed by the metadata with the files replaced by their fileref URIs. Thus, we need a way of determining if a particular output is a File that needs ingesting or just a metadata String that happens to be a gs url. I've had some luck in the past using womtool to describe the WDL and then walk the description of the outputs, dispatching to the relevant handlers (see workflows.clj in the test tools). For lack of a better alternative, I'm going to pursue using womtool to handle outputs in a generic way.

Implemented by changing the source and executors to be queues of `[description value]` pairs. In the case of the TDR sources, the description is a qualified keyword describing what the object is (ie. a snapshot). For the terra executor, the description is the workflow description as reported by womtool and the value is the workflow.